Make settings modal to use own labels
