+++ src/renderer/modules/AppModule/AppStarting/constants.ts
@@ -4,9 +4,7 @@ import { LanguageCode } from 'SettingsModule/types';
 import { LANGUAGE } from 'SettingsModule/constants';
 
 export interface Labels {
-  selectLanguage: string;
   language: string;
-  pleaseSelectLanguage: string;
   systemSettings: string;
   createAdmin: string;
   beforeStarting: string;
@@ -19,13 +17,6 @@ export interface Labels {
   noPasswordMatchTooltip: string;
   back: string;
   agree: string;
-  uiSize: string;
-  uiSizeMedium: string;
-  uiSizeLarge: string;
-  uiSizeExtraLarge: string;
-  colorTheme: string;
-  colorThemeDefault: string;
-  colorThemeDark: string;
   username: string;
   usernameTooltip: string;
   password: string;
@@ -44,7 +35,6 @@ export const labels: Record<LanguageCode, Labels> = {
   [LANGUAGE.KOREAN.languageCode]: korean
 };
 
-export const SELECT_LANGUAGE = 'selectLanguage' as const;
 export const SYSTEM_SETTINGS = 'systemSettings' as const;
 export const CREATE_ADMIN = 'createAdmin' as const;
 export const BEFORE_STARTING = 'beforeStarting' as const;
