+++ src/renderer/modules/AppModule/AppStarting/english.json
@@ -1,7 +1,5 @@
 {
   "language": "Language",
-  "pleaseSelectLanguage": "Please select a language",
-  "selectLanguage": "Select Language",
   "systemSettings": "System Settings",
   "createAdmin": "Create Admin",
   "beforeStarting": "Before Starting",
@@ -15,13 +13,6 @@
   "noConcentTooltip": "Please check 'I agree' on license agreement",
   "back": "Back",
   "agree": "Agree",
-  "uiSize": "UI Size",
-  "uiSizeMedium": "Medium",
-  "uiSizeLarge": "Large",
-  "uiSizeExtraLarge": "Extra Large",
-  "colorTheme": "Color Theme",
-  "colorThemeDefault": "Bright",
-  "colorThemeDark": "Dark",
   "username": "Username",
   "usernameTooltip": "You can't change the username of admin account",
   "password": "Password",
