+++ src/renderer/modules/AppModule/AppStarting/korean.json
@@ -1,7 +1,5 @@
 {
   "language": "언어",
-  "pleaseSelectLanguage": "언어를 선택하세요",
-  "selectLanguage": "언어 선택",
   "systemSettings": "환경 설정",
   "createAdmin": "관리자 계정 생성",
   "beforeStarting": "시작하기 전에",
@@ -14,13 +12,6 @@
   "noPasswordMatchTooltip": "다음 단계로 넘어가려면 모든 항목을 기입해 주세요",
   "back": "뒤로",
   "agree": "동의",
-  "uiSize": "크기",
-  "uiSizeMedium": "보통",
-  "uiSizeLarge": "큼",
-  "uiSizeExtraLarge": "매우 큼",
-  "colorTheme": "색상",
-  "colorThemeDefault": "밝게",
-  "colorThemeDark": "어둡게",
   "username": "사용자 이름",
   "usernameTooltip": "관리자 계정의 사용자 이름은 변경할 수 없습니다",
   "password": "비밀번호",
